test: move mocking logic into dagger modules
