+++ app/src/androidTest/java/com/chesire/malime/injection/modules/MockServerModule.kt
@@ -5,13 +5,15 @@ import com.chesire.malime.core.api.AuthApi
 import com.chesire.malime.core.api.Authorizer
 import com.chesire.malime.core.api.LibraryApi
 import com.chesire.malime.core.api.SearchApi
+import com.chesire.malime.core.flags.ItemType
 import com.chesire.malime.core.flags.SupportedService
+import com.chesire.malime.core.flags.UserSeriesStatus
 import com.chesire.malime.core.models.AuthModel
+import com.chesire.malime.core.models.MalimeModel
 import com.chesire.malime.core.repositories.Authorization
-import com.chesire.malime.mocks.MockLibraryApi
-import com.chesire.malime.mocks.MockSearchApi
 import dagger.Module
 import dagger.Provides
+import io.reactivex.Observable
 import io.reactivex.Single
 import javax.inject.Singleton
 
@@ -27,7 +29,7 @@ class MockServerModule {
 
     @Provides
     fun providesAuthApi(): AuthApi {
-        return (object : AuthApi {
+        return object : AuthApi {
             override fun login(username: String, password: String): Single<AuthModel> {
                 TODO("injected mock should be used")
             }
@@ -39,18 +41,49 @@ class MockServerModule {
             override fun getUserId(): Single<Int> {
                 TODO("injected mock should be used")
             }
-        })
+        }
     }
 
     @Provides
-    fun providesMalimeApi(mockApi: MockLibraryApi): LibraryApi = mockApi
+    fun providesLibraryApi(): LibraryApi {
+        return object : LibraryApi {
+            override fun addItem(item: MalimeModel): Single<MalimeModel> {
+                TODO("injected mock should be used") //To change body of created functions use File | Settings | File Templates.
+            }
+
+            override fun deleteItem(item: MalimeModel): Single<MalimeModel> {
+                TODO("injected mock should be used") //To change body of created functions use File | Settings | File Templates.
+            }
+
+            override fun getUserLibrary(): Observable<List<MalimeModel>> {
+                TODO("injected mock should be used") //To change body of created functions use File | Settings | File Templates.
+            }
+
+            override fun updateItem(
+                item: MalimeModel,
+                newProgress: Int,
+                newStatus: UserSeriesStatus
+            ): Single<MalimeModel> {
+                TODO("injected mock should be used") //To change body of created functions use File | Settings | File Templates.
+            }
+        }
+    }
 
     @Provides
-    fun providesSearchApi(mockApi: MockSearchApi): SearchApi = mockApi
+    fun providesSearchApi(): SearchApi {
+        return object : SearchApi {
+            override fun searchForSeriesWith(
+                title: String,
+                type: ItemType
+            ): Single<List<MalimeModel>> {
+                TODO("injected mock should be used")
+            }
+        }
+    }
 
     @Provides
     fun providesAuthorizer(): Authorizer<*> {
-        return (object : Authorizer<Int> {
+        return object : Authorizer<Int> {
             override fun storeAuthDetails(model: AuthModel) {
                 TODO("injected mock should be used")
             }
@@ -77,6 +110,6 @@ class MockServerModule {
             override fun clear() {
                 TODO("injected mock should be used")
             }
-        })
+        }
     }
 }
